backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-environ==0.11.2
djangorestframework==3.15.2
idna==3.10
psycopg2==2.9.10
requests==2.32.3
resend==2.4.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uuid==1.30
